Task2/requirements.txt
numpy
matplotlib
Pillow
opencv-python
shapely
ultralytics
